Added support for the FileChooser.setOpenFilesInPlace() method, which allows files to be opened in the gallery on the simulator in place rather than making a temp copy of the file. https://github.com/shannah/cn1-filechooser/issues/1 This is required for this commit to work in the CN1FileChooser project. https://github.com/shannah/cn1-filechooser/commit/8e94fa3412cb8e6571673fe4490c99d027c8b081

+++ Ports/JavaSE/src/com/codename1/impl/javase/JavaSEPort.java
@@ -8447,9 +8447,13 @@ public class JavaSEPort extends CodenameOneImplementation {
                             } else {
                                 ext= imageTypes[0];
                             }
-                            File tmp = File.createTempFile("temp", "." + ext);
-                            tmp.deleteOnExit();
-                            copyFile(selected, tmp);
+                            File tmp = selected;
+                            if (!"true".equals(Display.getInstance().getProperty("openGallery.openFilesInPlace", "false"))) {
+                                tmp = File.createTempFile("temp", "." + ext);
+                                tmp.deleteOnExit();
+                                copyFile(selected, tmp);
+                            }
+                            
                             result = new com.codename1.ui.events.ActionEvent("file://" + tmp.getAbsolutePath().replace('\\', '/'));
                         } catch(IOException err) {
                             err.printStackTrace();
